feat: move Settingsfragment to settings folder
